update: deploy createdAt time to use Heroku release's `updated_at`
Releases are created by Heroku, then the release phase runs, which
typically completes in a matter of seconds and updates the `update_at`
field.

We want to use `updated_at` instead of `created_at` so the release time
is marked when the release phase finished, rather than when it started.