add(sentry): configure sentry releases & upload sourcemaps
We compile and minify the code via esbuild which without sourcemaps gives
garbage stack traces in Sentry.

So this change sets up releases and uploads the sourcemaps for a given
release.

Also substitute in the release env var at build time, which is easier
than updating an env var in lambda.